server/project/requirements.txt
Django==1.11
django-auditlog==0.4.5
django-extra-fields==0.9
django-filter==1.1.0
django-jsonfield==1.0.1
djangorestframework==3.7.7
psycopg2-binary==2.7.4
python-dateutil==2.6.0
pytz==2018.3
six==1.11.0
